sd_data_and_training/requirements.txt
ultralytics>=8.3.0
roboflow>=1.1.40
opencv-python>=4.10.0
numpy>=1.26.0
PyYAML>=6.0.0
tqdm>=4.66.0
pandas>=2.2.0
